+++ nonebot_plugin_xiuxian_2/xiuxian/xiuxian_config.py
@@ -45,7 +45,7 @@ def convert_rank(rank_name):
     
 class XiuConfig:
     def __init__(self):
-        self.sql_table = ["xiuxian_user", "xiuxian_cooldown", "xiuxian_sect", "xiuxian_buff", "xiuxian_back", "xiuxian_impart"]  
+        self.sql_table = ["xiuxian_user", "xiuxian_time", "xiuxian_sect", "xiuxian_buff", "xiuxian_back", "xiuxian_impart"]  
         self.sql_user = ["id", "user_id", "user_name", "stone", "root",
                                  "root_type", "level", "power",
                                  "create_time", "is_sign", "is_beg", "is_ban",
@@ -54,7 +54,7 @@ class XiuConfig:
                                  "sect_position", "hp", "mp", "atk",
                                  "atkpractice", "sect_task", "sect_contribution",
                                  "sect_elixir_get", "blessed_spot_flag", "blessed_spot_name", "user_stamina"]
-        self.sql_cooldown = ["user_id", "type", "create_time", "scheduled_time", "last_check_info_time"]
+        self.sql_time = ["user_id", "type", "create_time", "scheduled_time", "last_check_info_time"]
         self.sql_sect = ["sect_id", "sect_name", "sect_owner", "sect_scale", "sect_used_stone", "sect_fairyland",
                           "sect_materials", "mainbuff", "secbuff", "elixir_room_level"]
         self.sql_buff = ["id", "user_id", "main_buff", "sec_buff", "faqi_buff", "fabao_weapon", "armor_buff",
